fix test imports
