+++ src/fides/cli/commands/util.py
@@ -9,7 +9,6 @@ from fides.cli.utils import (
     FIDES_ASCII_ART,
     check_and_update_analytics_config,
     check_server,
-    create_config_file,
     print_divider,
     send_init_analytics,
     with_analytics,
@@ -24,6 +23,7 @@ from fides.ctl.core.deploy import (
     teardown_application,
 )
 from fides.ctl.core.utils import echo_green
+from fides.ctl.core.config.helpers import create_config_file
 
 
 @click.command()
@@ -46,7 +46,7 @@ def init(ctx: click.Context, fides_directory_location: str) -> None:
 
     # create the config file as needed
     config_path = create_config_file(
-        ctx=ctx, fides_directory_location=fides_directory_location
+        config=config, fides_directory_location=fides_directory_location
     )
     print_divider()
