+++ src/fides/ctl/core/config/utils.py
@@ -1,11 +1,5 @@
 import logging
-import os
 from os import getenv
-from pathlib import Path
-from typing import Any, Dict, List, Union
-
-from click import echo
-from toml import dump, load
 
 from fides.ctl.core.utils import echo_red
